apps/blueprints-cli/requirements.txt
click==8.1.3
Jinja2==3.1.2
MarkupSafe==2.1.2
PyYAML==6.0
